Revert the no-op Float64(...) wrapper on _default_step_size's literals
957bfaaf wrapped 1e-3 and 1e-2 in Float64(...) "to match the funnel they
feed", but both literals are already Float64 -- the wrapper changes
nothing and is noise. The step-size funnel is deliberately Float64 by
design (the plan says so); the real fix in that commit was T(step_size)
at the _leaf_optim_step! call site, which stays as it was.

+++ src/optimizers/optimizer.jl
@@ -198,8 +198,8 @@ mutable struct Optimizer{MT <: GeometricOptimizers.OptimizerMethod, CT, ST, RT, 
     iterations::Int
 end
 
-_default_step_size(::GeometricOptimizers.Adam) = Float64(1e-3)
-_default_step_size(::GeometricOptimizers.OptimizerMethod) = Float64(1e-2)
+_default_step_size(::GeometricOptimizers.Adam) = 1e-3
+_default_step_size(::GeometricOptimizers.OptimizerMethod) = 1e-2
 
 _step_size(η::Real, ::Int) = Float64(η)
 # `t` and not `t - 1`: `optimization_step!` increments before it asks, so the first step of a solve
